Create GARCH Analysis SP500
GARCH model best fit created for period of 2008-2009.
The conditional standard error was calculated to create a confidence interval for predicted volatility in time series.

Peaks and troughs were mapped with consecutive periods of volatilities above threshold conditional standard deviation of 2.378 identified by cluster and duration of cluster.

Period of highest volatilities and dates were printed and identified on a chart. These were identified 3/10/2008 – 16/10/2008.

This period had the greatest volatility and was above the conditional standard deviation threshold of 2.378 which was significant at the 5% level. This period exhibited the steepest decline in SP500 price decline during the 2008 – 2009 financial crisis plummeting over 18%.